+++ src/app/cards/MarkdownTogglableDetails.tsx
